UML Diagram and JavaDoc Comments added; ReadMe partially updated.
I added JavaDoc comments to each class and generated JavaDoc documentation. I am also providing the UML Diagram for the project. I added to the ReadMe file, but I haven't finished the rest of the testing instructions. On a side note, I slightly modified the FXML files in order to make the interfaces of a shorter height.